greatly reduce memory usage of paths
Technically, this replaces a wavelet tree representing the path with packed
vector for ids and a marker bitvector to mark unlinked nodes. The memory
savings are significant, around 30% for simple linear graphs. It should be
noted that the linking and unlinking code still hasn't been tested.

This does remove a potentially important capability, which is to query the path
by nodes using wt_str.select(). However, this information is indexed already in
the graph in other wavelet trees, and so it isn't clear why we'd need that.