use a single wavelet tree to record mutable paths
The wavelet tree allows us to record all the paths and provides query of
occurrences by handle. As much of the cost in the WT is due to alphabet size,
this provides a substantial improvement where many paths overlap. This opens us
up to improvements in the underlying WT implementation, such as using a wavelet
matrix backed by a bitvector. The path_t type is removed and the whole
implementation is backed into graph_t.