store node layout in separate variables to improve performance
For path-free graphs, this comes at some cost. But when we have a number of
paths, it's essential, because we don't waste lots of time changing the node
header, which is at the front of our node byte array.

Also, we don't have to pack and unpack the layout anymore!