lazily destroy path steps
Previously, we had a lot of problems rewriting segments of paths and
destroying steps. The basic issue is that steps are identified by their
rank on a node, so removing a step earlier in the order will break
references to other steps on that node. A simple solution is to do what
was done for handles, and to mark the steps as deleted. Deleted steps
are skipped during iteration for each step on handle, and are by
definition not part of any path and thus unreachable during path
iteration. This makes path steps stable for all operations, at the cost
of requiring some kind of garbage collection to remove the deleted steps.

Removing the deleted steps requires a call to optimize, or a
sort/rebuild of the graph. As this would frequently be done following
modification of the graph that would delete steps, I don't think it will
cause any practical issues. The deleted steps, like handles, do not get
output in GFA or in any visualization or rendering of the graph.

To implement this efficiently, I have reserved the path handle id 0 to
mean "deleted step" internally. This generates an ABI incompatibility
with previous versions. For simplicity, this change is implemented at
the level of the graph itself, rather than using a hack inside of the
node object.