major rework of node_t, edge and path step storage semantics
Previously, we were storing path steps as deltas. This was efficient
when the graph is sorted. Packing and unpacking these relativistic
encodings is not trivial, so in this branch the idea is to do away with
these operations, replacing them with per-node alphabet encoder that
project the local neighborhood of nodes reached by edges and path steps
into a small alphabet.

Path steps now record on special bits if they are the start and/or end
of the given path.

The current branch does not pass tests. It does not yet have the
encoder, so memory usage will be significantly higher for large graphs.