Don't prevent non-x86 builds
This is pretty questionable. Presumably @ekg had a reason for putting
this error in, in commit 36fd016, and it probably has to do with the
lock-free implementation assuming TSO or something. But the file doesn't
use x86 assembly or anything, so I'd like to be able to build stuff on
ARM even if parts of it don't work yet so we can debug.